:recycle: Move token specs into own file
